backend/requirements.txt
# Django Core
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
django-extensions==3.2.0

# Database
psycopg2-binary==2.9.6
django-tenants==3.8.0

# Authentication
djangorestframework-simplejwt==5.2.2

# Environment & Configuration
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.0

# Multi-tenancy & Permissions
django-tenant-schemas==1.10.0
django-guardian==2.4.0

# Rate Limiting & Security
django-ratelimit==4.0.0
cryptography==41.0.3

# File Storage
django-storages==1.13.2
boto3==1.26.137
Pillow==9.5.0

# Caching & Background Tasks
redis==4.5.4
django-redis==5.2.0
celery==5.2.7

# AI & ML
openai==0.27.8
google-cloud-vision==3.4.4
pytesseract==0.3.10

# Web3
web3==6.0.0
eth-account==0.8.0

# Development & Testing
pytest==7.3.1
pytest-django==4.5.2
coverage==7.2.3

# Production
gunicorn==20.1.0
whitenoise==6.4.0
